Changes made to customer
