+++ src/main/java/com/vehicleRental/factories/CustomerFactory.java
@@ -1,20 +1,26 @@
 package com.vehicleRental.factories;
 
-import com.vehicleRental.domain.Address;
 import com.vehicleRental.domain.Customer;
+
+import java.util.Map;
 
 /**
  * Created by Noor on 27/10/2017.
  */
 public class CustomerFactory
 {
-    public static Customer getCustomer(String name, String surname, String email, Address address)
+    public static Customer getCustomer(Map<String, String> stringValues, Map<String,Integer> intValues)
     {
         Customer customer = new Customer.Builder()
-                .name(name)
-                .surname(surname)
-                .email(email)
-                .address(address)
+                .name(stringValues.get("name"))
+                .surname(stringValues.get("surname"))
+                .email(stringValues.get("email"))
+                .city(stringValues.get("city"))
+                .province(stringValues.get("province"))
+                .complex(stringValues.get("complex"))
+                .street(stringValues.get("street"))
+                .houseNumber(intValues.get("houseNumber"))
+                .postalCode(intValues.get("houseNumber"))
                 .build();
         return customer;
     }
