--- src/main/java/com/vehicleRental/repositories/AddressRepository.java
@@ -1,10 +1,0 @@
-package com.vehicleRental.repositories;
-
-import com.vehicleRental.domain.Address;
-import org.springframework.data.repository.CrudRepository;
-
-public interface AddressRepository extends CrudRepository<Address, Long> {
-    Address findAddressById(Long id);
-    Address findAddressByCustomer(Long id);
-
-}
